requirements.txt
streamlit
streamlit_option_menu
pandas
pillow
plotly